slight refactor synapse/abs workspace runner tests
